requirements-docker.txt
cffi==1.17.1
colorama==0.4.6
delta-spark==2.4.0
h11==0.16.0
idna==3.10
packaging==25.0
pandas==2.0.3
psutil==7.0.0
py4j==0.10.9.7
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sortedcontainers==2.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
websocket-client==1.8.0